GET SearchApplication Response
